backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Tasks Assíncronas
celery==5.3.6
redis==5.0.1

# Scraping & HTTP
playwright==1.41.0
beautifulsoup4==4.12.3
httpx==0.26.0

# APIs Externas
openai==1.10.0
twilio==8.11.1

# Utilidades
python-slugify==8.0.1
pillow==10.2.0
python-dateutil==2.8.2
structlog==24.1.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.0
flake8==7.0.0
mypy==1.8.0
